asyncRequest: Work around issue #10
asyncRequest assumed that the query function will always be asynchronous
(that the result will happen in the next event loop iteration). This is
obviously not the case - we can pass a `pure`, for example.

A carefully placed `yieldAff` ensures that it really happens in next
loop iteration, thereby again masking the real bug.

Of course the correct solution would be to fix the actual bug, but this
is good enough for now.